chore(module): Refactor module module
